enforce pascal case on space and operation names if the naming conventions flag is on
